Fixed created but empty files as well as some reformatting and cleanup

+++ vcell-util/src/main/java/org/vcell/util/recording/FileRecord.java
@@ -1,32 +1,34 @@
 package org.vcell.util.recording;
 
 import java.io.File;
+import java.io.Writer;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
 
 public class FileRecord extends Record {
     protected File logFile;
-    protected FileWriter fileWriter;
     protected StringBuffer sBuff;
 
     public FileRecord(String filePath){
         this.sBuff = new StringBuffer();
         this.logFile = new File(filePath);
-        this.fileWriter = null;
         try {
-        if ((logFile.exists() || logFile.createNewFile()) && !logFile.isDirectory()){
-            logFile.delete();
-            this.fileWriter = new FileWriter(logFile, false);
-        }
-            
+            if (logFile.exists() && !logFile.isDirectory()){
+                this.write("Test Write");
+                Files.delete(logFile.toPath());
+            }   
         } catch (IOException e){
-            throw new RuntimeException("File \"" + filePath + "\"  could not be created or opened.", e);
+            throw new RuntimeException("File \"" + filePath + "\"  could not be created, writen to, or opened.", e);
         }
     }
 
-    // Expose the underlying write method.
     public void print(String message) throws IOException {
+        this.buffer(message);
+        this.flush();
+    }
+
+    public void buffer(String message){
         this.sBuff.append(message);
     }
 
@@ -45,14 +47,17 @@ public class FileRecord extends Record {
     @Override
     public void close() throws IOException {
         this.flush();
-        this.fileWriter.close();
-        if (this.sBuff.toString().equals("")) 
+        if (this.sBuff.toString().equals("") && logFile.exists() && !logFile.isDirectory()) 
             Files.delete(this.logFile.toPath()); // Throws an exception on failure
     }
 
     @Override
     protected void write(String m) throws IOException {
-        if (!m.equals(""))
-            this.fileWriter.write(m);
+        if (logFile.isDirectory() || m.equals("")) 
+            return;
+
+        Writer fileWriter = new FileWriter(logFile, false);
+        fileWriter.write(m);
+        fileWriter.close();
     }   
 }
